WIP: abstract bot code
