+++ src/index.js
@@ -1,12 +1,10 @@
 const captains = console;
 const tmi = require('tmi.js');
-const fetch = require('node-fetch');
 const Discord = require('discord.js');
 const server = require('./server');
+const bot = require('./bot');
 
 server.start();
-
-captains.log(`Overlay ${process.env.greenscreenOverlayIframe}`);
 
 const channels = process.env.ttvChannels.toString().split(',');
 const clientUsername = process.env.clientUsername.toString();
@@ -40,15 +38,11 @@ const logger = require('./logger')(discordHook);
 
 discordHook.send('Client is online and running...');
 
-let conversationId;
-let conversationToken;
-let expiration;
-const azureBotToken = process.env.AzureBotToken;
 let moderators = [clientUsername];
 let isChatClientEnabled = true;
 let lightCommandUsed = '';
 
-createNewBotConversation();
+bot.createNewBotConversation();
 
 // #region chat client code
 
@@ -135,7 +129,8 @@ function parseChat(message, userName) {
         server.triggerSpecialEffect(commandMessage, userName);
       }
       server.updateOverlay(commandMessage);
-      return sendCommand(commandMessage, userName)
+      return bot
+        .sendCommand(commandMessage, userName)
         .then(result => {
           logger('info', `Successfully sent the command from ${userName}`);
           return result;
@@ -149,7 +144,7 @@ function parseChat(message, userName) {
   }
 
   if (isStreamElements(userName) && isSpecialEffectCommand(message)) {
-    return triggerEffect(message, userName);
+    return bot.triggerEffect(message, userName);
   }
 
   return Promise.resolve('there was nothing to do');
@@ -168,96 +163,6 @@ function isLightControlCommand(message) {
     lightCommandUsed = message.startsWith(command.toLowerCase());
     return !!lightCommandUsed;
   });
-}
-
-// #endregion
-
-// #region overlay update code
-
-// #endregion
-
-// #region bot code
-
-function createNewBotConversation() {
-  captains.log(`Starting a new bot conversation at: ${new Date()}`);
-  startBotConversation().then(result => {
-    // eslint-disable-next-line prefer-destructuring
-    conversationId = result.conversationId;
-    conversationToken = result.token;
-    const expiresIn = parseInt(result.expires_in, 10);
-    expiration = new Date().getSeconds() + expiresIn - 30;
-    createTimeout(expiration);
-  });
-}
-
-function createTimeout(expirationTime) {
-  const timeInMilliseconds = expirationTime * 1000;
-  setTimeout(createNewBotConversation, timeInMilliseconds);
-}
-
-function startBotConversation() {
-  const url = 'https://directline.botframework.com/api/conversations';
-  return fetch(url, {
-    method: 'POST',
-    headers: {
-      Authorization: `Bearer ${azureBotToken}`
-    }
-  })
-    .then(response => response.json())
-    .catch(error => {
-      captains.log(
-        `There was an error starting a conversation with the bot: ${error}`
-      );
-      return error;
-    });
-}
-
-function sendCommand(commandMessage, user) {
-  const fullMessage = { text: commandMessage, from: user };
-  const url = `https://directline.botframework.com/api/conversations/${conversationId}/messages`;
-  return fetch(url, {
-    method: 'POST',
-    headers: {
-      Authorization: `Bearer ${conversationToken}`,
-      'content-type': 'application/json'
-    },
-    body: JSON.stringify(fullMessage),
-    mode: 'cors'
-  })
-    .then(response => response)
-    .catch(error => {
-      captains.log(error);
-      return error;
-    });
-}
-
-function triggerEffect(message, userName) {
-  let effect;
-  if (
-    message.includes('subscribed') ||
-    message.includes('cheered') ||
-    message.includes('tipped')
-  ) {
-    effect = 'trigger new subscriber';
-  } else {
-    effect = 'trigger new follower';
-  }
-
-  if (effect) {
-    return sendCommand(effect, userName)
-      .then(result => {
-        captains.log(
-          `Successfully triggered ${effect} command from ${userName}`
-        );
-        return result;
-      })
-      .catch(error => {
-        captains.log(error);
-        return error;
-      });
-  }
-  captains.warn(`Unsupported effect was received in the message: ${message}`);
-  return Promise.resolve('TODO = put something useful here');
 }
 
 // #endregion
